requirements.txt
colorama
discord.py
python-dotenv
asyncio
